Migrating attest to cobra, moving public-key impl to folder
