+++ cmd/cosign/cli/options/predicate.go
@@ -1,0 +1,35 @@
+//
+// Copyright 2021 The Sigstore Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package options
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// PredicateOptions is the wrapper for predicate related options.
+type PredicateOptions struct {
+	Path string
+	Type string
+}
+
+// AddPredicateOptions adds the predicate related options to cmd.
+func AddPredicateOptions(cmd *cobra.Command, o *PredicateOptions) {
+	cmd.Flags().StringVar(&o.Path, "predicate", "",
+		"path to the predicate file.")
+
+	cmd.Flags().StringVar(&o.Path, "type", "custom",
+		"specify predicate type (default: custom) (slsaprovenance|link|spdx)")
+}
