+++ cmd/cosign/cli/publickey/public_key_test.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cli
+package publickey
 
 import (
 	"bytes"
